requirements.txt
discord.py
asyncio
gTTS
PyNaCl
rstr
shutil
